Refactor project into app package
