+++ app/main.py
@@ -1,0 +1,16 @@
+from __future__ import annotations
+import sys
+from PySide6.QtWidgets import QApplication
+from .ui.main_window import MainWindow
+
+
+def main():
+        app = QApplication(sys.argv)
+        app.setApplicationName("IRR Scheduler")
+        win = MainWindow()
+        win.show()
+        sys.exit(app.exec())
+
+
+if __name__ == "__main__":
+        main()
